mc_database/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
insightface>=0.7.3
onnxruntime>=1.16.0
faiss-cpu>=1.7.4
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
